pyproject.toml
[tool.poetry]
name = "codingtrackerserver"
version = "0.1.0"
description = "Server that goes with the codingTracker app. It use aiohttp and sqlite3.0 to save data from the client. It also provide an entry point to get a complete rundown of every language used."
authors = ["Emmanuel Guefif"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.5"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pytest = "^7.4.2"
pytest-aiohttp = "^1.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
